Inference/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

pathlib2>=2.3.0

python-dotenv>=1.0.0

# optional
# requests>=2.28.0
# docker>=6.0.0
# pytest>=7.0.0